pull out write assignments into separate file
